cloud-function/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage==1.15.0
xgboost==0.82
pandas==0.24.2
gunicorn
flask